+++ cmd/ttn-lw-stack/commands/start.go
@@ -163,10 +163,7 @@ var startCommand = &cobra.Command{
 				return shared.ErrInitializeIdentityServer.WithCause(err)
 			}
 			if config.Cache.Service == "redis" {
-				is.SetRedisCache(redis.New(&redis.Config{
-					Redis:     config.Cache.Redis,
-					Namespace: []string{"is", "cache"},
-				}))
+				is.SetRedisCache(redis.New(config.Cache.Redis.WithNamespace("is", "cache")))
 			}
 			if oauthMount := config.IS.OAuth.UI.MountPath(); oauthMount != "/" {
 				rootRedirect = web.Redirect("/", http.StatusFound, oauthMount)
@@ -175,10 +172,9 @@ var startCommand = &cobra.Command{
 
 		if start.GatewayServer || startDefault {
 			logger.Info("Setting up Gateway Server")
-			config.GS.Stats = &gsredis.GatewayConnectionStatsRegistry{Redis: redis.New(&redis.Config{
-				Redis:     config.Cache.Redis,
-				Namespace: []string{"gs", "connection", "stats"},
-			})}
+			config.GS.Stats = &gsredis.GatewayConnectionStatsRegistry{
+				Redis: redis.New(config.Cache.Redis.WithNamespace("gs", "connection", "stats")),
+			}
 			gs, err := gatewayserver.New(c, &config.GS)
 			if err != nil {
 				return shared.ErrInitializeGatewayServer.WithCause(err)
@@ -190,18 +186,17 @@ var startCommand = &cobra.Command{
 			redisConsumerGroup := "ns"
 
 			logger.Info("Setting up Network Server")
-			config.NS.ApplicationUplinks = nsredis.NewApplicationUplinkQueue(redis.New(&redis.Config{
-				Redis:     config.Redis,
-				Namespace: []string{"ns", "application-uplinks"},
-			}), 100, redisConsumerGroup, redisConsumerID)
-			config.NS.Devices = &nsredis.DeviceRegistry{Redis: redis.New(&redis.Config{
-				Redis:     config.Redis,
-				Namespace: []string{"ns", "devices"},
-			})}
-			nsDownlinkTasks := nsredis.NewDownlinkTaskQueue(redis.New(&redis.Config{
-				Redis:     config.Redis,
-				Namespace: []string{"ns", "tasks"},
-			}), 100000, redisConsumerGroup, redisConsumerID)
+			config.NS.ApplicationUplinks = nsredis.NewApplicationUplinkQueue(
+				redis.New(config.Redis.WithNamespace("ns", "application-uplinks")),
+				100, redisConsumerGroup, redisConsumerID,
+			)
+			config.NS.Devices = &nsredis.DeviceRegistry{
+				Redis: redis.New(config.Redis.WithNamespace("ns", "devices")),
+			}
+			nsDownlinkTasks := nsredis.NewDownlinkTaskQueue(
+				redis.New(config.Redis.WithNamespace("ns", "tasks")),
+				100000, redisConsumerGroup, redisConsumerID,
+			)
 			if err := nsDownlinkTasks.Init(); err != nil {
 				return shared.ErrInitializeNetworkServer.WithCause(err)
 			}
@@ -215,27 +210,22 @@ var startCommand = &cobra.Command{
 
 		if start.ApplicationServer || startDefault {
 			logger.Info("Setting up Application Server")
-			config.AS.Links = &asredis.LinkRegistry{Redis: redis.New(&redis.Config{
-				Redis:     config.Redis,
-				Namespace: []string{"as", "links"},
-			})}
-			config.AS.Devices = &asredis.DeviceRegistry{Redis: redis.New(&redis.Config{
-				Redis:     config.Redis,
-				Namespace: []string{"as", "devices"},
-			})}
-			config.AS.PubSub.Registry = &asiopsredis.PubSubRegistry{Redis: redis.New(&redis.Config{
-				Redis:     config.Redis,
-				Namespace: []string{"as", "io", "pubsub"},
-			})}
-			config.AS.ApplicationPackages.Registry = &asioapredis.ApplicationPackagesRegistry{Redis: redis.New(&redis.Config{
-				Redis:     config.Redis,
-				Namespace: []string{"as", "io", "applicationpackages"},
-			})}
+			config.AS.Links = &asredis.LinkRegistry{
+				Redis: redis.New(config.Redis.WithNamespace("as", "links")),
+			}
+			config.AS.Devices = &asredis.DeviceRegistry{
+				Redis: redis.New(config.Redis.WithNamespace("as", "devices")),
+			}
+			config.AS.PubSub.Registry = &asiopsredis.PubSubRegistry{
+				Redis: redis.New(config.Redis.WithNamespace("as", "io", "pubsub")),
+			}
+			config.AS.ApplicationPackages.Registry = &asioapredis.ApplicationPackagesRegistry{
+				Redis: redis.New(config.Redis.WithNamespace("as", "io", "applicationpackages")),
+			}
 			if config.AS.Webhooks.Target != "" {
-				config.AS.Webhooks.Registry = &asiowebredis.WebhookRegistry{Redis: redis.New(&redis.Config{
-					Redis:     config.Redis,
-					Namespace: []string{"as", "io", "webhooks"},
-				})}
+				config.AS.Webhooks.Registry = &asiowebredis.WebhookRegistry{
+					Redis: redis.New(config.Redis.WithNamespace("as", "io", "webhooks")),
+				}
 			}
 			as, err := applicationserver.New(c, &config.AS)
 			if err != nil {
@@ -246,14 +236,12 @@ var startCommand = &cobra.Command{
 
 		if start.JoinServer || startDefault {
 			logger.Info("Setting up Join Server")
-			config.JS.Devices = &jsredis.DeviceRegistry{Redis: redis.New(&redis.Config{
-				Redis:     config.Redis,
-				Namespace: []string{"js", "devices"},
-			})}
-			config.JS.Keys = &jsredis.KeyRegistry{Redis: redis.New(&redis.Config{
-				Redis:     config.Redis,
-				Namespace: []string{"js", "keys"},
-			})}
+			config.JS.Devices = &jsredis.DeviceRegistry{
+				Redis: redis.New(config.Redis.WithNamespace("js", "devices")),
+			}
+			config.JS.Keys = &jsredis.KeyRegistry{
+				Redis: redis.New(config.Redis.WithNamespace("js", "keys")),
+			}
 			js, err := joinserver.New(c, &config.JS)
 			if err != nil {
 				return shared.ErrInitializeJoinServer.WithCause(err)
@@ -329,10 +317,9 @@ var startCommand = &cobra.Command{
 
 		if start.DeviceClaimingServer || startDefault {
 			logger.Info("Setting up Device Claiming Server")
-			config.DCS.AuthorizedApplications = &dcsredis.AuthorizedApplicationRegistry{Redis: redis.New(&redis.Config{
-				Redis:     config.Redis,
-				Namespace: []string{"dcs", "applications", "authorized"},
-			})}
+			config.DCS.AuthorizedApplications = &dcsredis.AuthorizedApplicationRegistry{
+				Redis: redis.New(config.Redis.WithNamespace("dcs", "applications", "authorized")),
+			}
 			dcs, err := deviceclaimingserver.New(c, &config.DCS)
 			if err != nil {
 				return shared.ErrInitializeDeviceClaimingServer.WithCause(err)