+++ pkg/gatewayserver/redis/registry.go
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2020 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,9 +19,10 @@ import (
 
 	ttnredis "go.thethings.network/lorawan-stack/pkg/redis"
 	"go.thethings.network/lorawan-stack/pkg/ttnpb"
+	"go.thethings.network/lorawan-stack/pkg/unique"
 )
 
-// GatewayConnectionStatsRegistry implements the GatewayConnectionStatsRegistry interface
+// GatewayConnectionStatsRegistry implements the GatewayConnectionStatsRegistry interface.
 type GatewayConnectionStatsRegistry struct {
 	Redis *ttnredis.Client
 }
@@ -30,8 +31,9 @@ func (r *GatewayConnectionStatsRegistry) key(uid string) string {
 	return r.Redis.Key("uid", uid)
 }
 
-// Set sets or clears the connection stats for a gateway
-func (r *GatewayConnectionStatsRegistry) Set(ctx context.Context, uid string, stats *ttnpb.GatewayConnectionStats) error {
+// Set sets or clears the connection stats for a gateway.
+func (r *GatewayConnectionStatsRegistry) Set(ctx context.Context, ids ttnpb.GatewayIdentifiers, stats *ttnpb.GatewayConnectionStats) error {
+	uid := unique.ID(ctx, ids)
 	if stats == nil {
 		return r.Redis.Del(r.key(uid)).Err()
 	}
@@ -40,8 +42,9 @@ func (r *GatewayConnectionStatsRegistry) Set(ctx context.Context, uid string, st
 	return err
 }
 
-// Get returns the connection stats for a gateway
-func (r *GatewayConnectionStatsRegistry) Get(ctx context.Context, uid string) (*ttnpb.GatewayConnectionStats, error) {
+// Get returns the connection stats for a gateway.
+func (r *GatewayConnectionStatsRegistry) Get(ctx context.Context, ids ttnpb.GatewayIdentifiers) (*ttnpb.GatewayConnectionStats, error) {
+	uid := unique.ID(ctx, ids)
 	stats := &ttnpb.GatewayConnectionStats{}
 	err := ttnredis.GetProto(r.Redis, r.key(uid)).ScanProto(stats)
 	if err != nil {